back/requirements.txt
asn1crypto==0.24.0
astroid==2.0.4
autopep8==1.4.3
certifi==2018.10.15
cffi==1.11.5
chardet==3.0.4
Click==7.0
cryptography==2.4.2
Flask==1.0.2
idna==2.7
isort==4.3.4
itsdangerous==1.1.0
Jinja2==2.10
lazy-object-proxy==1.3.1
MarkupSafe==1.1.0
mccabe==0.6.1
nba-api==1.0.6
peewee==3.7.1
pycodestyle==2.4.0
pycparser==2.19
pylint==2.1.1
PyMySQL==0.9.2
python-dateutil==2.7.5
requests==2.20.1
six==1.11.0
urllib3==1.24.1
Werkzeug==0.14.1
wrapt==1.10.11
